+++ Changelog.md
@@ -14,6 +14,11 @@ every new version is a new major version.
 -   Exported `getWaitForDevice` hance allowing apps to resort previous state if
     needed
 
+### Added
+
+-   Flash Messages feature. Comes with convenience thunk functions in order to
+    create info, success, warning, and error messages.
+
 ### Changed
 
 -   FeedbackPane: Change **Platform** to **Operating system**, and remove the